+++ src/GenerativeAI/Models/ModelBase.cs
@@ -5,6 +5,10 @@ using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using GenerativeAI.Exceptions;
+using System.Net.Http.Headers;
+using System.Text.Json.Nodes;
+using System.Reflection;
+using System.Net;
 
 namespace GenerativeAI.Models
 {
@@ -20,7 +24,7 @@ namespace GenerativeAI.Models
             set => Client.Timeout = value;
         }
 
-        protected JsonSerializerOptions SerializerOptions => new JsonSerializerOptions()
+        protected JsonSerializerOptions SerializerOptions => new()
         {
             PropertyNamingPolicy = JsonNamingPolicy.CamelCase, Converters = { new JsonStringEnumConverter() },
             DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull
@@ -57,8 +61,45 @@ namespace GenerativeAI.Models
                 throw new GenerativeAIException($"Error while requesting {url.ToString("__API_Key__")}:\r\n\r\n{content}",content);
             }
         }
-               
 
+        protected virtual async IAsyncEnumerable<string> StreamContentAsync(string apiKey, string model, GenerateContentRequest request2,CancellationToken cancellationToken = default)
+        {
+            
+            var url = new RequestUrl(model, Tasks.StreamGenerateContent, apiKey, false, BaseUrl, Version);
+
+            var ms = new MemoryStream();
+            await JsonSerializer.SerializeAsync(ms, request2, SerializerOptions, cancellationToken);
+            ms.Seek(0, SeekOrigin.Begin);
+
+            using var request = new HttpRequestMessage(HttpMethod.Post, url);
+            request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
+            using var requestContent = new StreamContent(ms);
+            request.Content = requestContent;
+            requestContent.Headers.ContentType = new MediaTypeHeaderValue("application/json");
+            using var response = await Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead,cancellationToken).ConfigureAwait(false);
+            if (response.IsSuccessStatusCode)
+            {
+                using (var responseStream = await response.Content.ReadAsStreamAsync())
+                using (var streamReader = new StreamReader(responseStream))
+                {
+                    string line = string.Empty;
+                    while ((line = await streamReader.ReadLineAsync()) != null)
+                    {
+                        if (line.Contains(@"""text"""))
+                        {
+                            var jsonString = "{" + line + "}";
+                            var jsonObject = JsonSerializer.Deserialize<JsonObject>(jsonString);
+                            yield return jsonObject?["text"]?.ToString();
+                        }
+                    }
+                }
+            }
+            else
+            {
+                var content = await response.Content.ReadAsStringAsync();
+                throw new GenerativeAIException($"Error while requesting {url.ToString("__API_Key__")}:\r\n\r\n{content}", content);
+            }
+        }
         protected virtual async Task<CountTokensResponse> CountTokens(string apiKey, string model,
             CountTokensRequest request)
         {
@@ -71,7 +112,6 @@ namespace GenerativeAI.Models
             var response = await Client.PostAsync(url, stringContent);
             if (response.IsSuccessStatusCode)
             {
-                var str = await response.Content.ReadAsStringAsync();
                 var result = await JsonSerializer.DeserializeAsync(await response.Content.ReadAsStreamAsync(),
                     GoogleSerializerContext.Default.CountTokensResponse);
                 return result;