+++ tests/GenerativeAI.Tests/Model/GeminiProVision_Tests.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using GenerativeAI.Classes;
 using GenerativeAI.Models;
 using GenerativeAI.Types;
 using Xunit.Abstractions;
@@ -45,6 +40,68 @@ namespace GenerativeAI.Tests.Model
             var result = await visionModel.GenerateContentAsync(parts);
 
             Console.WriteLine(result.Text());
+        }
+
+        [Fact]
+        public async Task ShouldIdentifyImageWithFileObject()
+        {
+            var imageBytes = await File.ReadAllBytesAsync("image.png");
+
+            string prompt = "What is in the image?";
+
+            var apiKey = Environment.GetEnvironmentVariable("Gemini_API_Key", EnvironmentVariableTarget.User);
+
+            var visionModel = new GeminiProVision(apiKey);
+
+            var result = await visionModel.GenerateContentAsync(prompt,new FileObject(imageBytes,"image.png"));
+
+            Console.WriteLine(result.Text());
+        }
+
+        [Fact]
+        public async Task ShouldIdentifyImageWithFileObjectWithStreaming()
+        {
+            var imageBytes = await File.ReadAllBytesAsync("image.png");
+
+            string prompt = "What is in the image?";
+
+            var apiKey = Environment.GetEnvironmentVariable("Gemini_API_Key", EnvironmentVariableTarget.User);
+
+            var visionModel = new GeminiProVision(apiKey);
+
+            Action<string> handler = (a) =>
+            {
+                Console.WriteLine(a);
+            };
+
+            var result = await visionModel.StreamContentAsync(prompt, new FileObject(imageBytes, "image.png"),handler);
+
+            Console.WriteLine("");
+            Console.WriteLine(result);
+        }
+
+        [Fact]
+        public async Task ShouldIdentifyImageWithFileObjectWithChatAndStreaming()
+        {
+            var imageBytes = await File.ReadAllBytesAsync("image.png");
+
+            string prompt = "What is in the image?";
+
+            var apiKey = Environment.GetEnvironmentVariable("Gemini_API_Key", EnvironmentVariableTarget.User);
+
+            var visionModel = new GeminiProVision(apiKey);
+
+            var chat = visionModel.StartChat(new StartChatParams());
+
+            Action<string> handler = (a) =>
+            {
+                Console.WriteLine(a);
+            };
+
+            var result = await chat.StreamContentVisionAsync(prompt, new FileObject(imageBytes, "image.png"), handler);
+
+            Console.WriteLine("");
+            Console.WriteLine(result);
         }
     }
 }