+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-### Version 1.12
+### Version 0.1.14
+    Added Streaming Support
+    Added ModelInfoService to get the available generative AI models.
+    Simplified Gemini Pro Vision Generation with new FileObject class
+    
+### Version 0.1.12
     Added Gemini Pro Vision Model support
     Added GenerativeAIException for more details of error
 
-### Version 1.11 and Below
+### Version 0.1.11 and Below
     Initial SDK Implementaton